+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Data Engineers Snowflake DataOps Utils Project Changelog
 This file contains the changelog for the Data Engineers Snowflake DataOps Utils project, detailing updates, fixes, and enhancements made to the project over time.
 
+## v0.3.10.1 - 2026-02-12 - Grant Application Select
+
+* added macro `grant_object_application` to apply the required permissions on the provided object to an application
+* updated macro `grant_schema_read` to only process `ROLE` and not `APPLICATION` grant statements
+
 ## v0.3.10 - 2026-02-10 - Grant Shares
 
 * added macro `grant_internal_share_read` to apply `select` permissions on all tables and views